requirements.txt
websockets
pika
python-dotenv
psycopg2-binary
dbt-postgres
duckdb
pandas
sqlalchemy
